同步原项目代码

+++ src/Form/Field/MultipleFile.php
@@ -4,7 +4,6 @@ namespace Encore\Admin\Form\Field;
 
 use Encore\Admin\Form\Field;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\Validator;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class MultipleFile extends Field
@@ -81,7 +80,7 @@ class MultipleFile extends Field
 
         list($rules, $input) = $this->hydrateFiles(Arr::get($input, $this->column, []));
 
-        return Validator::make($input, $rules, $this->validationMessages, $attributes);
+        return \validator($input, $rules, $this->getValidationMessages(), $attributes);
     }
 
     /**
@@ -218,10 +217,12 @@ class MultipleFile extends Field
         $config = [];
 
         foreach ($files as $index => $file) {
-            $config[] = [
+            $preview = array_merge([
                 'caption' => basename($file),
                 'key'     => $index,
-            ];
+            ], $this->guessPreviewType($file));
+
+            $config[] = $preview;
         }
 
         return $config;
@@ -240,8 +241,7 @@ class MultipleFile extends Field
     }
 
     /**
-     * @param array $options
-     * @param array $options
+     * @param string $options
      */
     protected function setupScripts($options)
     {
@@ -329,7 +329,9 @@ EOT;
     /**
      * Destroy original files.
      *
-     * @return string.
+     * @param string $key
+     *
+     * @return array.
      */
     public function destroy($key)
     {
@@ -337,7 +339,7 @@ EOT;
 
         $file = Arr::get($files, $key);
 
-        if ($this->storage->exists($file)) {
+        if (!$this->retainable && $this->storage->exists($file)) {
             $this->storage->delete($file);
         }
